separeted user module
