+++ base/templates/base/main.html
@@ -35,7 +35,7 @@
         }
 
         .container {
-            max-width: 550px;
+            max-width: 1000px;
             margin: auto;
             background-color: #fff;
             -webkit-box-shadow: 2px 2px 13px -4px rgba(0, 0, 0, 0.21);
