+++ base/templates/base/task_list.html
@@ -23,10 +23,6 @@
     <a id="add-link" href="{% url 'task-create' %}">&#x2b; Add</a>
 </div>
 
-
-
-
-
 <div class="task-items-wrapper">
     {% for task in tasks %}
         <div class="task-wrapper">
@@ -35,14 +31,22 @@
                     <div class="task-complete-icon"></div>
                     <i><s><a href="{% url 'task-update' task.id %}">{{task}}</a></s></i>
                 </div>
-                <a class ="delete-link" href="{% url 'task-delete' task.id %}">&#215;</a>
+                
+                <!--<a class ="delete-link" href="{% url 'task-delete' task.id %}">&#215;</a>-->
                 {% else %}
                 <div class="task-title">
                     <div class="task-incomplete-icon"></div>
                     <a href="{% url 'task-update' task.id %}">{{task}}</a>
                 </div>
+              
+                {% endif %}
+                <div>
+                    <p>Assigned: {{user}}</p>
+                </div>
+                <div>
+                    <p>Status: {{task.status}}</p>
+                </div>
                 <a class ="delete-link" href="{% url 'task-delete' task.id %}">&#215;</a>
-            {% endif %}
             
         </div>
