+++ base/views.py
@@ -14,31 +14,6 @@ from django.contrib.auth import login
 
 from .models import Task 
 
-class CustomLoginView(LoginView):
-   template_name = 'base/login.html'
-   fields = '__all__'
-   redirect_authenticated_user = True
-
-   def get_success_url(self):
-      return reverse_lazy('tasks')
-   
-class RegisterPage(FormView):
-   template_name = 'base/register.html'
-   form_class = UserCreationForm
-   redirect_authenticated_user = True
-   success_url = reverse_lazy('tasks')
-
-   def form_valid(self, form):
-      user = form.save()
-      if user is not None:
-         login(self.request, user)
-      return super(RegisterPage, self).form_valid(form)
-   
-   def get(self, *args, **kwargs):
-      if self.request.user.is_authenticated:
-         return redirect('task')
-      return super(RegisterPage, self).get(*args, **kwargs)
-
 class TaskList(LoginRequiredMixin, ListView):
    model = Task
    context_object_name = 'tasks'
