+++ taskmanager/example_settings.py
@@ -43,6 +43,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'base.apps.BaseConfig',
+    'users.apps.UsersConfig',
 ]
 
 MIDDLEWARE = [
