+++ users/templates/users/register.html
@@ -6,6 +6,7 @@
 </div>
 
 <div class="card-body">
+    <h1>This is new page</h1>
     <form method="POST">
         {% csrf_token %}
         <label>{{form.username.label}}</label>
